// desktop/src/open-api/model/currencySeparator.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type CurrencySeparator = ',' | '.';

export const CurrencySeparator = {
    Comma: ',' as CurrencySeparator,
    Period: '.' as CurrencySeparator
};